file_api_project/requirements.txt
Django==5.1.4
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1